Add docstring for replication monitoring script
